+++ src/angular-ui-handsontable.js
@@ -341,15 +341,17 @@ angular.module('uiHandsontable', [])
             , $container = uiDatagrid.$container;
 
           var isSelected
-          , lastSelectionRow
-          , lastSelectionCol;
+            , lastSelectionRow
+            , lastSelectionCol;
 
           $container.on('selection.handsontable', function (event, r, c, r2, c2) {
             isSelected = true;
             lastSelectionRow = r;
             lastSelectionCol = c;
 
-            if (!scope.$$phase && typeof scope.selectedIndex !== 'undefined' && scope.selectedIndex != r) { //if digest is not in progress
+            if (!scope.$$phase && typeof scope.selectedIndex === 'object' && typeof scope.selectedIndex !== 'undefined' && scope.selectedIndex != r) {
+              //make sure digest is not in progress
+              //make sure selectgedIndex is observable to avoid "Non-assignable model expression" error
               scope.$apply(function () {
                 scope.selectedIndex = r;
               });
@@ -360,7 +362,9 @@ angular.module('uiHandsontable', [])
             isSelected = false;
             lastSelectionRow = null;
 
-            if (!scope.$$phase && typeof scope.selectedIndex !== 'undefined' && scope.selectedIndex !== null) { //if digest is not in progress
+            if (!scope.$$phase && typeof scope.selectedIndex === 'object' && typeof scope.selectedIndex !== 'undefined' && scope.selectedIndex != null) {
+              //make sure digest is not in progress
+              //make sure selectgedIndex is observable to avoid "Non-assignable model expression" error
               scope.$apply(function () {
                 scope.selectedIndex = null;
               });